Move the batched-FFT and row-major-LDE unit tests into corresponding file
